+++ src/main/java/com/github/jlangch/venice/impl/types/collections/VncHashSet.java
@@ -177,15 +177,13 @@ public class VncHashSet extends VncSet {
 		}
 		else if (Types.isVncHashSet(o)) {
 			final Integer sizeThis = size();
-			final Integer sizeOther = size();
+			final Integer sizeOther = ((VncHashSet)o).size();
 			int c = sizeThis.compareTo(sizeOther);
 			if (c != 0) {
 				return c;
 			}
 			else {
-				if (equals(o)) {
-					return 0;
-				}
+				return equals(o) ? 0 : -1;
 			}
 		}