+++ src/main/java/com/github/jlangch/venice/impl/types/collections/VncJavaMap.java
@@ -222,15 +222,13 @@ public class VncJavaMap extends VncMap implements IVncJavaObject {
 		}
 		else if (Types.isVncJavaMap(o)) {
 			final Integer sizeThis = size();
-			final Integer sizeOther = size();
+			final Integer sizeOther = ((VncJavaMap)o).size();
 			int c = sizeThis.compareTo(sizeOther);
 			if (c != 0) {
 				return c;
 			}
 			else {
-				if (equals(o)) {
-					return 0;
-				}
+				return equals(o) ? 0 : -1;
 			}
 		}