+++ src/main/java/com/github/jlangch/venice/impl/types/collections/VncJavaSet.java
@@ -186,15 +186,13 @@ public class VncJavaSet extends VncSet implements IVncJavaObject {
 		}
 		else if (Types.isVncJavaSet(o)) {
 			final Integer sizeThis = size();
-			final Integer sizeOther = size();
+			final Integer sizeOther = ((VncJavaSet)o).size();
 			int c = sizeThis.compareTo(sizeOther);
 			if (c != 0) {
 				return c;
 			}
 			else {
-				if (equals(o)) {
-					return 0;
-				}
+				return equals(o) ? 0 : -1;
 			}
 		}