+++ src/main/java/com/github/jlangch/venice/impl/types/collections/VncMutableMap.java
@@ -229,15 +229,13 @@ public class VncMutableMap extends VncMap {
 		}
 		else if (Types.isVncMutableMap(o)) {
 			final Integer sizeThis = size();
-			final Integer sizeOther = size();
+			final Integer sizeOther = ((VncMutableMap)o).size();
 			int c = sizeThis.compareTo(sizeOther);
 			if (c != 0) {
 				return c;
 			}
 			else {
-				if (equals(o)) {
-					return 0;
-				}
+				return equals(o) ? 0 : -1;
 			}
 		}