+++ src/main/java/com/github/jlangch/venice/impl/types/collections/VncOrderedMap.java
@@ -249,15 +249,13 @@ public class VncOrderedMap extends VncMap {
 		}
 		else if (Types.isVncOrderedMap(o)) {
 			final Integer sizeThis = size();
-			final Integer sizeOther = size();
+			final Integer sizeOther = ((VncOrderedMap)o).size();
 			int c = sizeThis.compareTo(sizeOther);
 			if (c != 0) {
 				return c;
 			}
 			else {
-				if (equals(o)) {
-					return 0;
-				}
+				return equals(o) ? 0 : -1;
 			}
 		}