+++ src/main/java/com/github/jlangch/venice/impl/types/collections/VncSortedMap.java
@@ -247,15 +247,13 @@ public class VncSortedMap extends VncMap {
 		}
 		else if (Types.isVncSortedMap(o)) {
 			final Integer sizeThis = size();
-			final Integer sizeOther = size();
+			final Integer sizeOther = ((VncSortedMap)o).size();
 			int c = sizeThis.compareTo(sizeOther);
 			if (c != 0) {
 				return c;
 			}
 			else {
-				if (equals(o)) {
-					return 0;
-				}
+				return equals(o) ? 0 : -1;
 			}
 		}