+++ src/main/java/com/github/jlangch/venice/impl/types/collections/VncSortedSet.java
@@ -169,15 +169,13 @@ public class VncSortedSet extends VncSet {
 		}
 		else if (Types.isVncSortedSet(o)) {
 			final Integer sizeThis = size();
-			final Integer sizeOther = size();
+			final Integer sizeOther = ((VncSortedSet)o).size();
 			int c = sizeThis.compareTo(sizeOther);
 			if (c != 0) {
 				return c;
 			}
 			else {
-				if (equals(o)) {
-					return 0;
-				}
+				return equals(o) ? 0 : -1;
 			}
 		}